requirements.txt
Django==4.0.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers 
psycopg2-binary 
python-decouple

# Outras dependências úteis
requests==2.27.1  # Para fazer requisições HTTP (opcional)
gunicorn==20.1.0  # Para rodar o servidor em produção (opcional)
